cephadm-preflight: handle client nodes
This adds the possibility to setup repositories on nodes intended to be
used as 'clients'.
Basically, it requires you to define a new group called "clients" in
your existing inventory.

eg:

 $ cat hosts
 mynode1
 mynode2
 mynode3

 [clients]
 client1
 client2
 client3

Then, you can run the playbook as documented:

$ ansible-playbook -i hosts cephadm-preflight.yml

This will install the packages 'chronyd' and 'ceph-common' on those
nodes.

Signed-off-by: Guillaume Abrioux <gabrioux@redhat.com>