clients: Add new client playbook
Playbook handles the distribution of conf and keyring
to a group of ceph clients. The expectation is that the
ceph admin has created at least the appropriate keyring
on the admin node for this playbook to distribute. If the
playbook is called without a conf, it will generate and
distribute a minimal conf

Signed-off-by: Paul Cuzner <pcuzner@redhat.com>