tox: use quay.io registry
We already configured the pacific branch to use the quay.io registry
and the ceph/ceph container images.
This has been done by 571cbb4 so the credentials were also updated to
use the one from quay.io instead of quay.ceph.io.
Unfortunatly the devel branch was still using quay.ceph.io which broke
the CI jobs on that branch.
Since we have all the daemon and daemon-base container images on quay.io
now then we can use that registry for all branches.
This also updates the contianer image tag to use the latest-master-devel
tag instead of latest-master as the devel tag is rebuild daily.

Signed-off-by: Dimitri Savineau <dsavinea@redhat.com>