preflight/clients: workaround packaging issue
After an upgrade from RHCS 4, we must be sure ceph-common-14.x is
uninstalled before installing that package from rhceph-5 repository.
Otherwise, it leads to dependencies errors like following:

```
    Depsolve Error occured:
     Problem: problem with installed package ceph-osd-2:14.2.11-184.el8cp.x86_64
      - package ceph-osd-2:14.2.11-184.el8cp.x86_64 requires ceph-base = 2:14.2.11-184.el8cp, but none of the providers can be installed
      - package ceph-osd-2:14.2.4-125.el8cp.x86_64 requires ceph-base = 2:14.2.4-125.el8cp, but none of the providers can be installed
```

This could be addressed by using the `allowerasing` option of the ansible module `package`
but this option is available in 2.10 onward only.

Closes: https://bugzilla.redhat.com/show_bug.cgi?id=2008402

Signed-off-by: Guillaume Abrioux <gabrioux@redhat.com>