Changed "distribution" to "os_family" to include Debian
Changed the `when` variable from `distribution` to `os_family` to
include Debian, Ubuntu and possibly other debian based distros too.
Instead of install podman or docker by distribution version, the
playbook now search for the podman package in the repositories, if the
package exists, install it, if not, install docker.