rpm: set Epoch to 1
This will override all usual version comparisons.

In RHCS 6, we shipped v3, and we want users to go back to v2. This epoch
bump will cause users to "upgrade" to the highest epoch version. We will
apply this change to RHCS 6+.

Apply it upstream as well so that we keep parity with downstream and
preserve the ability to upgrade from downstream -> upstream versions.